+++ index.html
@@ -80,5 +80,6 @@
 </footer>
 <script src="dist/bannerClick.bundle.js"></script>
 <script src="dist/formAbandonment.bundle.js"></script>
+<script src="dist/ecommerceTracking.bundle.js"></script>
 </body>
 </html>
