+++ src/modules/buildEcommerceObj.js
@@ -1,0 +1,31 @@
+import{calculateQuantity} from './calculateQuantity';
+import{buildProductsObjArray} from './buildProductsObjArray';
+import{calculateTotalPrice} from './calculateTotalPrice';
+import{removeDuplicates} from './removeDuplicates';
+import{extractMetaTagContent} from './extractMetaTagContent';
+import{extractProductName} from './extractProductName';
+import{buildProductIndex} from './buildProductIndex';
+
+let categoriesArray = extractMetaTagContent('WT\\.pn_fa', 'Categories meta tag not available').split(',');
+let productsArray = extractMetaTagContent('WT\\.pn_sku', 'Product meta tag not available').split(',');
+let pricesArray = extractMetaTagContent('WT\\.tx_s', 'Prices meta tag not available').split(',');
+
+export let buildEcommerceObj = {
+    'event' : 'checkoutOption',
+    'ecommerce' : {
+        'actionField' : {
+            'step' : extractMetaTagContent('WT\\.si_p', 'Meta tag not available'),
+            'option' : 'Receipt'
+        },
+        'purchase' : {
+            'actionField' : {
+                'id' : extractMetaTagContent('WT\\.tx_id', 'Meta tag not available'),
+                'affiliation' : extractMetaTagContent('WT\\.si_n', 'Meta tag not available'),
+                'revenue' : calculateTotalPrice(pricesArray),
+                'tax' : '0.00',
+                'shipping' : '0.00'
+            },
+            'products' : buildProductsObjArray(extractProductName(productsArray), removeDuplicates(buildProductIndex(productsArray, pricesArray, categoriesArray)), calculateQuantity(productsArray))
+        }
+    }
+};
