+++ src/modules/removeDuplicates.js
@@ -1,5 +1,5 @@
 export let removeDuplicates = function(array) {
     return array.filter(function (value, index, self) {
-        return self.indexOf(value) === index;           //this is simplified - need to flesh it out to understand
+        return self.indexOf(value) === index;
     });
 };
